api/requirements.txt
pytest
httpx
requests
fastapi
pydantic
uvicorn
python-dotenv
openai
langchain
ipykernel